api/requirements.txt
click==7.1.1
Flask==1.1.2
Flask-Cors==3.0.8
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
python-chess==0.30.1
python-dotenv==0.12.0
six==1.14.0
Werkzeug==1.0.1
